Mega WIP Update
64-bit
OpenXR (Khronos Loader)

+++ Projects/Android/jni/rtcw/src/qcommon/vm_local.h
@@ -134,7 +134,7 @@ struct vm_s {
 	// DO NOT MOVE OR CHANGE THESE WITHOUT CHANGING THE VM_OFFSET_* DEFINES
 	// USED BY THE ASM CODE
 	int programStack;               // the vm may be recursively entered
-	int ( *systemCall )( int *parms );
+	intptr_t ( *systemCall )( intptr_t *parms );
 
 	//------------------------------------
 
@@ -142,7 +142,7 @@ struct vm_s {
 
 	// for dynamic linked modules
 	void        *dllHandle;
-	int ( QDECL *entryPoint )( int callNum, ... );
+	intptr_t ( QDECL *entryPoint )( int callNum, ... );
 
 	// for interpreted modules
 	qboolean currentlyInterpreting;
@@ -180,5 +180,5 @@ int VM_CallInterpreted( vm_t *vm, int *args );
 vmSymbol_t *VM_ValueToFunctionSymbol( vm_t *vm, int value );
 int VM_SymbolToValue( vm_t *vm, const char *symbol );
 const char *VM_ValueToSymbol( vm_t *vm, int value );
-void VM_LogSyscalls( int *args );
+void VM_LogSyscalls( intptr_t *args );
